refactor: rename components folder from events with private routing convention
