+++ src/app/events/_components/event-card.tsx
