+++ src/app/events/_components/hero-section.tsx
